Refactored project structure to match other branches.
