+++ src/components/batches/Header/Header.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import Header from './Header';
+import Header from '.';
 
 const component = () => {
   return <Header />;
